+++ src/modules/sbstudio/plugin/model/__init__.py
@@ -10,7 +10,12 @@ from .pyro_control import PyroControlPanelProperties
 from .safety_check import SafetyCheckProperties, get_overlay as get_safety_check_overlay
 from .settings import DroneShowAddonFileSpecificSettings
 from .show import DroneShowAddonProperties
-from .storyboard import ScheduleOverride, StoryboardEntry, Storyboard
+from .storyboard import (
+    ScheduleOverride,
+    StoryboardEntry,
+    StoryboardEntryOrTransition,
+    Storyboard,
+)
 
 __all__ = (
     "DroneShowAddonFileSpecificSettings",
@@ -26,6 +31,7 @@ __all__ = (
     "SafetyCheckProperties",
     "ScheduleOverride",
     "StoryboardEntry",
+    "StoryboardEntryOrTransition",
     "Storyboard",
     "get_formation_order_overlay",
     "get_safety_check_overlay",