+++ src/modules/sbstudio/plugin/panels/light_effects.py
@@ -101,9 +101,32 @@ class LightEffectsPanel(Panel):
                     layout.separator()
 
             col = layout.column()
-            col.prop(entry, "frame_start")
-            col.prop(entry, "duration")
-            col.prop(entry, "frame_end")
+            # TODO: remove bullet from beginning of text on UI somehow
+            col.prop_search(
+                entry,
+                "storyboard_entry_or_transition_selection",
+                scene.skybrush.storyboard,
+                "entries_or_transitions",
+                text="Attach to",
+            )
+            if entry.storyboard_entry_or_transition_selection:
+                col.prop(
+                    entry,
+                    "frame_start",
+                    text=f"Start Frame ({entry.frame_start_offset:+})",
+                )
+                col.prop(
+                    entry,
+                    "duration",
+                    text=f"Duration ({entry.duration_offset:+})",
+                )
+                col.prop(
+                    entry, "frame_end", text=f"End Frame ({entry.frame_end_offset:+})"
+                )
+            else:
+                col.prop(entry, "frame_start")
+                col.prop(entry, "duration")
+                col.prop(entry, "frame_end")
             col.separator()
             col.prop(entry, "fade_in_duration")
             col.prop(entry, "fade_out_duration")